requirements.txt
Django==4.2.7
django-crispy-forms==2.0
crispy-tailwind==0.5.0
python-decouple==3.8
bleach==6.2.0
build==1.2.2.post1
django-filter==24.3
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
docutils==0.21.2
environ==1.0
folium==0.19.4
frozenlist==1.5.0
installer==0.7.0
instaloader==4.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
selenium-stealth==1.0.6
Pillow>=10.0
gunicorn==20.1.0  
cloudinary==1.39.0
django-cloudinary-storage==0.3.0

